chore: drops Transaction.ProcessRequest into an own package. (#296)
* chore: drops process request into an own package.

* fix: moves missing function.

* fix: makes lint happy.

* fix: seclang tests.
